RequestStop and HasStopped for ZmqSubscriber.
This extends the ZmqSubscriber functionality with a new method that
allows to just request it to stop, without yet joining the thread
(this in particular can be used from the ZMQ thread itself safely).

If the thread has just been requested to stop, stopped, but the
ZmqSubscriber itself has not been Stop()'ed yet, it can still be
restarted safely.

Also adds a new notification for ZmqListener's, which tells them when
the ZMQ listening thread actually stops.