Introduce DISCONNECTED GSP state.
This introduces DISCONNECTED as new potential state of the GSP.  When
the ZMQ subscriber is stopped, the state is changed to DISCONNECTED
(and also the state is now DISCONNECTED instead of UNKNOWN before the
GSP is started up properly).

For now, about the only situation in which this can happen is if
one of the ZmqListener methods throws an exception (e.g. because
the Xaya Core RPC has some issues), in which case the ZmqSubscriber now
stops itself.