Logic for probing & fixing the Xaya connection.
This implements Game::ProbeAndFixConnection.  It implements basic logic
for checking the staleness of the ZMQ connection.  If it is stale, then
the ZMQ subscriber is stopped (disconnecting the game in the follow-up).
When the game is disconnected on the other hand, it tries to reconnect
ZMQ and reinitialise the game state.

By default, the connection is considered stale after 120 seconds.  This is
typically long enough so that blocks will arrive well within; but even if
not, the logic tries to "ping" the ZMQ connection (by requesting a dummy
game_sendupdates) when half the staleness period is reached.

So it should (for a working connection) never happen that it gets
disconnected; but even if that might happen due to some edge-case effects,
that is also not a big deal, as it just will be reconnected shortly after.

For now, the new logic is not actually used, except in the unit test.
But we will introduce a background thread that will call the check and
reconnection repeatedly from a running Game as a follow-up.