Share common code for ZMQ connection.
This factors out the common code (between Game::Start() and the reconnection
attempt in ProbeAndFixConnection) for connecting the GSP daemon to the
ZMQ server and initialising the state.

This actually also fixes a bug, because the reconnection logic did not
have the code to disable pending tracking in the ZMQ server in case
no pending handler was installed (causing a potential assert failure).