Introduce libshipscore.
This splits libships (the Xayaships game logic) into libshipscore and
libships.  The former contains just the core logic needed on the channel,
without dependencies on e.g. SQLite or libgamechannel (just channelcore).

This is useful in attempts to build e.g. Xayaships in a browser, where the
core stuff can be compiled into wasm and used with libchannelcore.