Add sqliteintro, with table introspection.
The new file xayagame/sqliteintro is added, which will contain utilities
for introspection of an SQLite database, and the automatic hashing of the
full game state (for fork detection and debugging).

For now, it contains introspection for listing all tables, all
columns of a table, and the primary key of a table (for a deterministic
sorting order).