Properly handle errors in websocket script.
Catch exceptions in the websocket script (either from the RPC or the
websocket sending), and properly handle them.  When an error is caught,
the script waits for a configurable time, before trying again.  (Previously
the looping thread would just crash and the script continue running but
without being responsive at all.)