Basic framework for game-state processors.
The new framework around SQLiteProcessor's allows basic post-processing to
be done on SQLite game states, most importantly hashing for fork debugging.

For now the framework is very basic and all computation is done synchronously,
but in the future it can be changed to the processing is asynchronous
on a snapshot instead (and does not block the GSP process while running).