Rename targetBlockHash in Game.
This is a simple rename of Game::targetBlockHash to "catchingUpTarget".
This block hash is what we currently catch up to from attach notifications,
and the new name clarifies this.

The "target" block will be something else, namely a desired target to which
to sync to, and then stop (as an optional new feature being implemented).