Introduce "target block" concept in Game.
This adds a new functionality to Game:  A "target block" can be designated;
if so, then the game tries to sync to exactly this block, stopping as soon
as it is reached in state "AT_TARGET".  The block can be updated, to make the
game sync again to tip or another target.

Note that for now, setting to a target block in the past does not request
notifications to it (because Xaya X would not support that).  So it is
mainly useful to make the GSP stop at a future block, which is already
something that can be used well for debugging.