Update interface of SQLiteProcessor for async.
In preparation of supporting async processing of the game state, this updates
the interface of SQLiteProcessor so that it accepts (and gets passed from
SQLiteGame) an optional read-only DB snapshot.

For now it does not yet use it, but with these changes, it will be a
change purely internal inside of SQLiteProcessor to turn on async
processing where possible.