Explicitly close SQLite database in default main.
Explicitly invoke CloseDatabase() in the SQLite default main function
before it returns.  This ensures that the database is closed and all
GSP processes attached to it are cleaned up before the default main
returns.

Previously, the storage would be cleaned up only in the destructor of the
SQLiteGame subclass, which would run typically in the GSP main function
outside of the default main.

With this, we in particular ensure that Finish() will be called on all
active processors before the default main finishes, which is what users
would expect, and which makes sure that they don't have to use a particular
order for destruction between the game rules and processors.