Run SQLiteProcessor's async.
This implements asynchronous processing for SQLiteProcessor's:  If it is
possible to get a read-only snapshot of the database, then the actual
processing (in Compute()) is run on a separate thread using that
snapshot, so that the main GSP can continue processing blocks while
the processing is being done.  At a later time, when the processor is ready,
the result will be Store()'ed.