Define utility class for temporary file names.
We use temporary names for SQLite databases in some of the unit tests,
using std::tmpnam and removing with std::remove in the test fixture's
destructor.

This introduces a helper class TempFileName, which abstracts this behaviour,
so it can be more easily used (and used in more tests as well).