Request updates to explicit target block.
When an explicit target block is set, request game_sendupdates notifications
from Game with that target set as toblock argument.  This ensures that we
can properly sync to the target block, even if it is e.g. in the past.

For this to work if a GSP is connected to Xaya X rather than Xaya Core,
we need https://github.com/xaya/xayax/pull/8.