Allow disabling TLS verification in RestClient.
This adds a new method to the RestClient, which can be used to disable
TLS verification done on requests.  While this is obviously dangerous
(and opt-in), it can be useful in some situations, where the data is
validated additionally anyway (e.g. when it is used to load the preimage
of some hash that has been posted on-chain).  Using this option in this
situation allows to avoid any fussing around with CA files.