Integrate coprocessors with Game.
This integrates Coprocessors with Game:  Game now holds a batch of
coprocessors, to which external users can add their own if they like
with a new method.

When blocks are processed forward/backward or the initial state is
set, then Game ensures that the coprocessors in its batch are
updated, too.  Also, while a GameLogic callback is running, the
current coprocessor blocks can be accessed by user code through the Context.