Callback for changes to GSP instance state.
This adds a new notification callback to GameLogic, which gets invoked
not only when the game state changes, but the instance state of the GSP
(i.e. also when things like reconnect/disconnect, reaching target block
and so on happen).  The callback gets passed the instance state in the
form that GetCustomStateData returns it.

While the callback is executed, the main lock on Game is held; so the
caller might, if they want, also extract game-state data and pair it with
the instance state they get if needed.