Don't attempt SQLite snapshots if not up-to-date.
When the game is not up-to-date, do not attempt to create SQLite
snapshots for the state snapshot or coprocessors in the first place,
to avoid the extra IO overhead in these situations.

Snapshots are meant to be used to remove lock contention once the
GSP is fully synced, not during syncing.