Interpret new "error" response for game updates.
game_sendupdates may now return an "error" field, indicating if there was
some issue on the base-chain side (such as requested a "fromblock" that
Xaya X is not yet synced to).

This change makes the GSP interpret this field; if it is present and
set to true, then the ZMQ connection is reset, so that it gets restored
and the request retried later to recover.