No transaction logic in Coprocessor::Block
This removes the handling of atomic transactions (begin/commit/abort)
and RAII semantics for that out of Coprocessor::Block.  Instead, starting
and resolving transactions is now done through Coprocessor itself -- this
allows to decouple them from individual blocks, and e.g. do batched commits
while syncing (as we do with the underlying Storage already).

For now, the Game instance itself handles the begin/commit/abort for
the coprocessor transactions, doing one explicitly for each block as it
was done before.