Fix missing lock for GetCustomStateData.
In SQLiteGame::GetCustomStateData, when we can't use an existing state
snapshot, we need to get the state from the main database.  In this case,
the main Game lock must not be released before we have finished the
processing.

Thus we need to use the method variant that takes over the lock, and not
the one without that releases it immediately.